Add reveal hook and Projects split
Introduce a unified scroll-reveal system and refactor animations/styles across the site. Adds a useReveal hook and Reveal component and applies them to FeaturedProjects and Portfolio so headers and cards self-reveal (admin views stay static). Removes many per-component keyframe entrances in favor of staggered reveal using CSS vars (--reveal-index / --reveal-stagger) and consolidates transition timings to theme variables; tweaks durations for landing, contact, and admin media. Prevents drag/floating clones from replaying entrance animations. Split the /projects page: extract ProjectsBody into a new client component, render it inside a Suspense boundary with a skeleton fallback so the banner/filter paint immediately while the card grid streams. Overall improves load/animation consistency and preserves admin drag/edit ergonomics.
