+++ src/app/(index)/featured.module.scss
@@ -1,13 +1,6 @@
 @use "@/styles/variables" as *;
 @use "@/styles/reorder" as r;
 
-// Entrance animates `translate`/opacity (not `transform`) so it never collides
-// with the pointer-tilt transform applied on hover.
-@keyframes fadeInUp {
-  from { opacity: 0; translate: 0 20px; }
-  to { opacity: 1; translate: 0 0; }
-}
-
 .sectionHeader {
   font-size: var(--fs-title);
   padding-block: 0.5em;
@@ -50,13 +43,24 @@
   text-decoration: none;
 
   // `--ty` carries the hover lift inside the transform so it composes with the
-  // pointer tilt (--rx/--ry) instead of overwriting it.
+  // pointer tilt (--rx/--ry) instead of overwriting it. The scroll-reveal
+  // entrance (global `.reveal`) animates the separate `translate` property and
+  // opacity, so it never collides with this transform.
   --ty: 0px;
   transform: perspective(900px) rotateX(var(--rx, 0deg)) rotateY(var(--ry, 0deg)) translateY(var(--ty));
-  transition: transform 0.25s ease-out, box-shadow 0.3s ease-in-out;
-
-  opacity: 0;
-  animation: fadeInUp 0.6s ease-out forwards;
+  // One declaration so the reveal entrance (opacity/translate, staggered) and the
+  // hover feedback (transform/box-shadow, no delay) coexist — a separate `.reveal`
+  // transition shorthand would clobber this one (or vice-versa). The delay list
+  // lines up with the property list: entrance staggers, hover is immediate.
+  transition:
+    opacity var(--dur-slow) var(--ease-out),
+    translate var(--dur-slow) var(--ease-out),
+    transform var(--dur-base) var(--ease-out),
+    box-shadow var(--dur-base) var(--ease-standard);
+  transition-delay:
+    calc(var(--reveal-index, 0) * var(--reveal-stagger)),
+    calc(var(--reveal-index, 0) * var(--reveal-stagger)),
+    0s, 0s;
 
   &:hover {
     --ty: -5px;
@@ -102,7 +106,7 @@
   font-weight: bold;
   text-decoration: none;
   border-radius: var(--radius-sm);
-  transition: transform 0.3s ease-in-out, box-shadow 0.3s ease-in-out, background-color 0.2s ease-in-out;
+  transition: transform var(--dur-base) ease-in-out, box-shadow var(--dur-base) ease-in-out, background-color var(--dur-fast) ease-in-out;
 }
 
 // Hover/focus feedback on the button itself, now that it's the only link in the
@@ -119,10 +123,10 @@
 }
 
 .floating {
-  // Cancel the entrance animation, whose filled transform would otherwise
-  // override the tilt below. The base .card starts at opacity 0, so restore it.
-  animation: none;
+  // Admin drag clone — kept fully opaque and untransformed by entrance state
+  // (it never carries the `.reveal` class) so it tracks the cursor cleanly.
   opacity: 1;
+  translate: none;
   @include r.floating-card;
 }
 
@@ -138,7 +142,7 @@
   width: 100%;
   height: 100%;
   object-fit: cover;
-  transition: transform 0.3s ease-in-out;
+  transition: transform var(--dur-base) ease-in-out;
 }
 
 .body {
@@ -152,7 +156,7 @@
     font-size: 1.35rem;
     color: var(--text-primary);
     margin-bottom: 0.65rem;
-    transition: color 0.3s ease-in-out;
+    transition: color var(--dur-base) ease-in-out;
   }
 
   p {
