+++ src/app/admin/media/media.module.scss
@@ -422,7 +422,7 @@
   display: flex;
   flex-direction: column;
   box-shadow: 0 1px 2px rgba(15, 23, 42, 0.05);
-  transition: box-shadow 0.15s ease, transform 0.15s ease;
+  transition: box-shadow var(--dur-fast) ease, transform var(--dur-fast) ease;
 
   &:hover {
     box-shadow: 0 6px 18px rgba(15, 23, 42, 0.12);
@@ -464,7 +464,7 @@
     background: var(--bg-card);
     box-shadow: 0 1px 3px rgba(15, 23, 42, 0.25);
     cursor: pointer;
-    transition: background 0.12s ease, border-color 0.12s ease;
+    transition: background var(--dur-fast) ease, border-color var(--dur-fast) ease;
 
     &::after {
       content: "";
