+++ src/components/Banner/banner.module.scss
@@ -9,7 +9,7 @@
   );
   display: grid;
   place-items: center;
-  animation: pageFadeIn 0.5s ease-out both;
+  animation: pageFadeIn 0.8s ease-out both;
 }
 
 .container {
